Bug: Server throws error if any one client disonnects.
Reason: Windows doesn't give the client a chance to send one last final packet (with the FIN flag set to true) over to the server to indicate that the connection is finished. It will kill it completely, and then the server thinks something unexpected happened with that socket.

Fix: To do the same operations that we do in the 'end' event listener in an 'error' listener as well.